test: move shared fixture code to the common dir
Move the code that was shared between tests to a common directory.
Declare different setup function for each suite that reuse the shared code.
