+++ test/polygon/catalyst/assetAttributesRegistry/getAttributes.test.ts
@@ -1,15 +1,13 @@
-import {ethers, deployments} from 'hardhat';
+import {deployments, ethers} from 'hardhat';
 import {Address, Receipt} from 'hardhat-deploy/types';
 import {BigNumber, Contract, Event} from 'ethers';
 import {expect} from '../../../chai-setup';
 import catalysts from '../../../../data/catalysts';
 import gems from '../../../../data/gems';
-import {setupGemsAndCatalysts} from '../gemsCatalystsRegistry/fixtures';
-import {setupAssetAttributesRegistry} from '../assetAttributesRegistry/fixtures';
+import {setupAssetAttributesRegistry, setupGemsAndCatalysts} from './fixtures';
 import {MintOptions} from '../assetMinter/fixtures';
 import {findEvents} from '../../../utils';
-import {transferSand} from '../utils';
-import {prepareGemEventData, getReceiptObject} from '../utils';
+import {getReceiptObject, prepareGemEventData, transferSand} from '../utils';
 
 const NFT_SUPPLY = 1;
